@@ -3,7 +3,7 @@ package edu.cmu.pocketsphinx;
 import android.app.Service;
 import android.content.Intent;
 import android.os.IBinder;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.util.Log;
 
 import edu.cmu.pocketsphinx.kit.RecognizerSetupListener;
